Fixed infinite loops in provideChange and exactChangePossible
Fixed infinite loops in the provideChange and exactChangePossible methods. @Simran the provideChange method currently fails to provide the change. The if and else if statements within the for loop always evaluate to false. Please have another look at this method to make sure the proper one evaluates to true when change needs to be provided.